+++ src/engine/Math/vector.ts
@@ -187,14 +187,15 @@ export class Vector implements Clonable<Vector> {
    * @param magnitude
    */
   public clampMagnitude(magnitude: number): Vector {
-    const size = this.size;
+    const size = this.magnitude;
     const newSize = clamp(size, 0, magnitude);
-    this.size = newSize;
+    this.magnitude = newSize;
     return this;
   }
 
   /**
    * The size (magnitude) of the Vector
+   * @deprecated Will be removed in v1, use Vector.magnitude
    */
   public get size(): number {
     return this.distance();
@@ -203,10 +204,26 @@ export class Vector implements Clonable<Vector> {
   /**
    * Setting the size mutates the current vector
    * @warning Can be used to set the size of the vector, **be very careful using this, mutating vectors can cause hard to find bugs**
+   * @deprecated Will be removed in v1, use Vector.magnitude
    */
   public set size(newLength: number) {
     const v = this.normalize().scale(newLength);
     this.setTo(v.x, v.y);
+  }
+
+  /**
+   * The magnitude (length) of the Vector
+   */
+  public get magnitude(): number {
+    return this.distance();
+  }
+
+  /**
+   * Setting the size mutates the current vector
+   * @warning Can be used to set the size of the vector, **be very careful using this, mutating vectors can cause hard to find bugs**
+   */
+  public set magnitude(newMagnitude: number) {
+    this.normalize().scale(newMagnitude, this);
   }
 
   /**